careershield_ai/ml_backend/requirements.txt
pandas
numpy
scikit-learn
flask
flask-cors
google-generativeai
cloudpickle
gunicorn
python-dotenv
